Replace sledgehammer with regular sized hammer
Contrail interpolation fix now with at least 15% less hackery. No seg faults pls.
