+++ src/chimera/fix/contrail_fix.S
@@ -2,32 +2,49 @@
 
 .intel_syntax noprefix
 
+.data
+init_value: .long 0x3D088889
+
 .text
-    .globl _new_contrail_update_function
-    _new_contrail_update_function:
-        ;# Check if we can update
-        push eax
-        mov eax, dword ptr [_can_update_contrail]
-        test al, al
 
-        ;# If nonzero, we can update
-        jnz do_the_thing
-        pop eax
+.globl _new_contrail_update_function
+_new_contrail_update_function:
+    ;# Check if we can update
+    push eax
+    mov eax, dword ptr [_can_update_contrail]
+    test al, al
 
-        ;# Nope? Okay
-        ret
+    ;# If nonzero, we can update
+    jnz do_the_thing
+    pop eax
 
-        ;# Let's do this!
-        do_the_thing:
+    ;# Nope? Okay
+    ret
 
-        ;# First, move zero to our _can_update_contrail value
-        xor eax, eax
-        mov dword ptr [_can_update_contrail], eax
+    ;# Let's do this!
+    do_the_thing:
 
-        ;# Next, set value in the stack to 1 / effective tick rate
-        mov eax, dword ptr [_update_contrail_by]
-        mov [esp + 0x8], eax
+    ;# First, move zero to our _can_update_contrail value
+    xor eax, eax
+    mov dword ptr [_can_update_contrail], eax
 
-        ;# We're done. Update
-        pop eax
-        jmp dword ptr [_original_contrail_update_function]
+    ;# Next, set value in the stack to 1 / effective tick rate
+    mov eax, dword ptr [_update_contrail_by]
+    mov [esp + 0x8], eax
+
+    ;# We're done. Update
+    pop eax
+    jmp dword ptr [_original_contrail_update_function]
+
+;# Block contrail update for first tick of exitence to prevent artifacts
+.globl _interpolation_memes
+_interpolation_memes:
+    push eax
+    mov eax, dword ptr [init_value]
+    cmp eax, dword ptr [esi + 0x24]
+    pop eax
+    jg skip
+    jmp dword ptr [_original_instruction]
+
+    skip:
+    jmp dword ptr [_skip_update]
