+++ src/chimera/signature/signature.cpp
@@ -94,6 +94,8 @@ namespace Chimera {
         FIND("command_list_custom_edition_sig", "core_custom_edition", { 0xBB, -1, -1, -1, -1, 0xBD, -1, -1, -1, -1, 0x8B, 0xFF, 0x8B, 0x33, 0x8A, -1, 0x18, 0xE8 });
         FIND("contrail_update_sig", "client", { 0x83, 0xEC, 0x10, 0x53, 0x55, 0x8B, 0x2D, -1, -1, -1, -1, 0x57, 0x83, 0xCB, 0xFF, 0x8B, 0xD3, 0x8B, 0xFD });
         FIND("contrail_update_func_sig", "client", { 0xE8, -1, -1, -1, -1, 0x57, 0xE8, -1, -1, -1, -1, 0x83, 0xC4, 0x08, 0x85, 0xF6, 0x5F });
+        FIND("contrail_skip_sig", "client", { 0x8A, 0x42, 0x02, 0xD9, 0x44, 0x24, 0x14, 0xA8, 0x20, 0xD9, 0x42, 0x2C });
+        FIND("contrail_interpolation_fix_sig", "client", { 0x8B, 0x1D, -1, -1, -1, -1, 0x66, 0x8B, 0x4E, 0x0E, 0x8B, 0x5B, 0x34, 0x25, 0xFF, 0xFF, 0x00, 0x00 });
         FIND("motion_sensor_update_sig", "client", { 0xA1, -1, -1, -1, -1, 0xDB, 0x40, 0x0C, 0xD8 ,0x0D, -1, -1, -1, -1, 0xDD, 0x05, -1, -1, -1, -1 });
         FIND("tab_out_video_1_sig", "client", { 0x38, 0x1D, -1, -1, -1, -1, 0x74, 0x0E, 0x83, 0xFD, 0x01, 0x74, 0x09, 0x83, 0xFD, 0x02, 0x0F, 0x85 });
         FIND("tab_out_video_2_sig", "client", { 0x75, 0x11, 0x38, 0x1D, -1, -1, -1, -1, 0x75, 0x04, 0x3A, 0xC3, 0x74, 0x05, 0xC6, 0x44, 0x24, 0x17, 0x01 });
